requirements.txt
# MCP服务器核心依赖
mcp-server-fastmcp>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 异步网络请求
aiohttp>=3.8.0
httpx>=0.24.0

# AI集成
openai>=1.0.0
anthropic>=0.7.0

# 数据处理
pyyaml>=6.0
ujson>=5.7.0

# 代码分析
astroid>=3.0.0
mypy>=1.0.0
black>=23.0.0
isort>=5.12.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 日志和监控
structlog>=23.0.0
rich>=13.0.0

# 缓存
cachetools>=5.3.0

# 工具库
pathlib2>=2.3.0; python_version < "3.4"
typing-extensions>=4.5.0

# 可选依赖（用于配置文件）
# PyYAML>=6.0  # 已包含在上面

# 开发依赖
# flake8>=6.0.0
# mypy>=1.0.0
# pre-commit>=3.0.0



